+++ AZAppearanceKit/AZTabBar.h
@@ -23,7 +23,7 @@
 /** Specifies the toolbar shadow's opacity.
  
  By default is `0.5`. */
-@property (nonatomic, assign) float shadowOpacity;
+@property (nonatomic, assign) CGFloat shadowOpacity;
 
 /** Specifies the gradient's start color.
